prism-framework/requirements-auth.txt
pyjwt==2.8.0 